refactor(audio): lift the in-process ACE-Step runtime out of the backend
ace_step_backend.py carried two unrelated jobs at 900 lines: the
MusicGenerator adapter that turns a GenerationRequest into a caption and a
GenerationResult over one of two transports, and the host-level work of
loading ACE-Step 1.5 into this process — device and precision selection,
MLX and torch allocator policy, handler construction, and running one
generation on the worker thread.

The second comes away cleanly. It reads five scalar config fields and
nothing else on the backend, owns every optional-ML import in the file, and
its measured constants are postmortems that only make sense together: the
4 GiB MLX cache limit that stopped macOS killing the process at 100 GiB,
and the deliberate refusal to clamp ACESTEP_MLX_VAE_CHUNK. Those comments
moved verbatim.

ace_step_backend.py 900 -> 540; ace_step_runtime.py 420. The only lines
added to the shrinking file are its docstring, the import block, and the
build_v15_runtime call replacing the deleted body.

Symbols crossing the new boundary become public, matching memory_budget and
face_scoring. is_ace_step_importable is still imported into
ace_step_backend and looked up there, so its seven patch targets keep the
ace_step_backend module path — patching ace_step_runtime instead would not
affect the backend.

Refs #686
