+++ vulture-whitelist.py
@@ -77,8 +77,8 @@ _.aspect_ratio  # unused property (src/immich_memories/api/models.py:425)
 _.is_landscape  # unused property (src/immich_memories/api/models.py:451)
 __getattr__  # unused function (src/immich_memories/audio/__init__.py:84)
 __getattr__  # unused function (src/immich_memories/audio/generators/__init__.py:29)
-_validate_torchcodec  # unused function (src/immich_memories/audio/generators/ace_step_backend.py:134)
-_mood_to_ace_prompt  # unused function (src/immich_memories/audio/generators/ace_step_backend.py:334)
+_validate_torchcodec  # unused function (src/immich_memories/audio/generators/ace_step_runtime.py:135)
+_mood_to_ace_prompt  # unused function (src/immich_memories/audio/generators/ace_step_backend.py:80)
 _.generate_with_stems  # unused method (src/immich_memories/audio/generators/base.py:96)
 _.generate_with_stems  # unused method (src/immich_memories/audio/generators/musicgen_backend.py:115)
 _.to_search_params  # unused method (src/immich_memories/audio/mood_analyzer.py:80)
